Re-build Y carriage for slic3r
Cura is able to slice the original Y carriage, but both slic3r and slic3r-PE
fail to see it as a manifold.  I used FreeCAD to convert the iges file to
an stl, and then used OpenSCAD to import the stl and add the spacers back to
it, and now both slic3r variants can slice it.  I sliced it with slic3r-PE
to print on the X5S with PLAPro/PLA+ and have included that gcode.